+++ tests/retrieval/helpers.py
@@ -41,6 +41,7 @@ def _compute_sklearn_metric(
     indexes: np.ndarray = None,
     metric: Callable = None,
     empty_target_action: str = "skip",
+    reverse: bool = False,
     **kwargs
 ) -> Tensor:
     """ Compute metric with multiple iterations over every query predictions set. """
@@ -67,7 +68,7 @@ def _compute_sklearn_metric(
     for group in groups:
         trg, pds = target[group], preds[group]
 
-        if trg.sum() == 0:
+        if ((1 - trg) if reverse else trg).sum() == 0:
             if empty_target_action == 'skip':
                 pass
             elif empty_target_action == 'pos':
@@ -268,8 +269,9 @@ class RetrievalMetricTester(MetricTester):
         sk_metric: Callable,
         dist_sync_on_step: bool,
         metric_args: dict,
+        reverse: bool = False,
     ):
-        _sk_metric_adapted = partial(_compute_sklearn_metric, metric=sk_metric, **metric_args)
+        _sk_metric_adapted = partial(_compute_sklearn_metric, metric=sk_metric, reverse=reverse, **metric_args)
 
         super().run_class_metric_test(
             ddp=ddp,
@@ -290,10 +292,10 @@ class RetrievalMetricTester(MetricTester):
         metric_functional: Callable,
         sk_metric: Callable,
         metric_args: dict,
+        reverse: bool = False,
         **kwargs,
     ):
-        # action on functional version of IR metrics is to return `tensor(0.0)` if not target is positive.
-        _sk_metric_adapted = partial(_compute_sklearn_metric, metric=sk_metric, **metric_args)
+        _sk_metric_adapted = partial(_compute_sklearn_metric, metric=sk_metric, reverse=reverse, **metric_args)
 
         super().run_functional_metric_test(
             preds=preds,
@@ -313,7 +315,6 @@ class RetrievalMetricTester(MetricTester):
         metric_module: Metric,
         metric_functional: Callable,
     ):
-        # action on functional version of IR metrics is to return `tensor(0.0)` if not target is positive.
         def metric_functional_ignore_indexes(preds, target, indexes):
             return metric_functional(preds, target)
 
@@ -337,7 +338,6 @@ class RetrievalMetricTester(MetricTester):
         if not torch.cuda.is_available():
             pytest.skip()
 
-        # action on functional version of IR metrics is to return `tensor(0.0)` if not target is positive.
         def metric_functional_ignore_indexes(preds, target, indexes):
             return metric_functional(preds, target)